feat: add `registerViewMorphAPI`
